+++ messytables/html.py
@@ -4,6 +4,9 @@ from collections import defaultdict
 
 
 class HTMLTableSet(TableSet):
+    """
+    A TableSet from a HTML document.
+    """
     def __init__(self, fileobj=None, filename=None, window=None):
 
         if filename is not None:
@@ -31,6 +34,9 @@ class HTMLTableSet(TableSet):
                 "other tables. This is a bug."  # avoid infinite loops
 
     def make_tables(self):
+        """
+        Return a listing of tables (as HTMLRowSets) in the table set.
+        """
         def rowset_name(rowset, table_index):
             return "Table {0} of {1}".format(table_index + 1,
                                              len(self.htmltables))
@@ -54,6 +60,9 @@ def insert_blank_cells(row, blanks):
 
 
 class HTMLRowSet(RowSet):
+    """
+    A RowSet representing a HTML table.
+    """
     def __init__(self, name, sheet, window=None):
         self.name = name
         self.sheet = sheet
@@ -166,6 +175,7 @@ class HTMLCell(Cell):
     def properties(self):
         return HTMLProperties(self._lxml)
 
+
 def text_from_element(elem):
     r"""
     >>> x = lxml.html.fromstring('''<td><center><span style="display:none; speak:none" class="sortkey">01879-07-01-0000</span>
@@ -184,10 +194,11 @@ def text_from_element(elem):
         else:
             cell_str = (x.text or '') + (x.tail or '')
         cell_str = cell_str.replace('\n', ' ').strip()
-        if x.tag =='br' or x.tag == 'p':
+        if x.tag == 'br' or x.tag == 'p':
             cell_str = '\n' + cell_str
         builder.append(cell_str)
     return ''.join(builder).strip()
+
 
 def is_invisible_text(elem):
     flag = False
@@ -197,6 +208,7 @@ def is_invisible_text(elem):
                 flag = True
 
     return flag
+
 
 class HTMLProperties(CoreProperties):
     KEYS = ['_lxml', 'html', 'colspan', 'rowspan']