+++ CHANGELOG.md
@@ -1,5 +1,13 @@
-0.13.0 (wip, master) (3 May 2014)
+0.14.1 (wip, master) (3 May 2014)
 * #25 Ignore invisible text in HTML cells
+
+0.14.0 (wip, master)
+* Add null processor
+* Update magic library
+
+0.13.0 (15 October 2013)
+* #86 Provide more information about the original table (properties)
+* #89 Skip PDF tests if PDFTables not installed
 
 0.12.0 (7 August 2013)
 * #51 Removed FloatType, use DecimalType